Ge-update zodat het werkt met floats ipv int

+++ opdracht 1/bol.py
@@ -1,4 +1,4 @@
 import math
-n, R = [int(x) for x in input().split()]
+n, R = [float(x) for x in input().split()]
 volume = (math.pow(math.pi, n/2)/math.factorial(n/2))*math.pow(R,n)
 print(volume)
